Make NodeInfo module available all MRML nodes in the Python console as n.NodeName
For example, the node named `MRHead` becomes available in the Python console as `n.MRHead`.

Type `n.` in the Python console and hit `Tab` key to see the list of all nodes (that are not hidden).
Characters in the node name that are not allowed in Python attribute names are converted to underscore.
Node names starting with a number are converted to variable names using the `n` prefix.

See https://discourse.slicer.org/t/make-it-easier-to-select-input-nodes-for-python-scripts/22498 for related discussion.